nexus-api/iiot_gateway_project/services/data_processing/requirements.txt
fastapi
uvicorn[standard]
pydantic
# asyncpg>=0.24.0,<0.25.0
aioredis
influxdb-client
numpy
python-multipart
# Tambahkan yang lain sesuai kebutuhan